pyproject.toml
[project]
name = "actbots"
version = "0.1.0"
description = "A clean, reusable Python library that lets developers spin up AI agents whose reasoning loops automatically *search → load → execute Jentic workflows and API operations."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "jentic>=0.8.1",
    "litellm>=1.0.0",
    "mem0ai>=0.1.114",
    "mypy>=1.16.1",
    "openai>=1.0",
    "pydantic>=2.0",
    "pytest>=8.0",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
    "ruff>=0.12.2",
]

[tool.actbots]
[tool.actbots.logging]

[tool.actbots.logging.console]
enabled = true
level = "INFO"
colored = true
format = "%(name)s:%(levelname)s: %(message)s"

[tool.actbots.logging.file]
enabled = true
level = "DEBUG"
path = "jentic_agents/logs/actbots.log"
format = "%(asctime)s - %(levelname)-8s - %(name)s - %(message)s"

[tool.actbots.logging.file.rotation]
enabled = true
max_bytes = 10485760
backup_count = 5

[tool.actbots.logging.loggers]
jentic_agents = { level = "DEBUG" }
chromadb = { level = "WARNING" }
httpx = { level = "WARNING" }
LiteLLM = { level = "WARNING" }
openai = { level = "WARNING" }

[tool.actbots.llm]
provider = "gemini"
model = "gemini-2.5-flash"

[tool.actbots.memory]
chroma_path = "jentic_agents/data/chroma_db"
llm_provider = "gemini"
llm_model = "gemini-2.5-flash"
embed_model = "text-embedding-3-small"

[tool.actbots.discord]
enabled = true
token = ""
target_user_id = 1343280955263160424
monitored_channels = [1391788053424509028]
default_channel_id = 1391788053424509028
escalation_channel_id = 1391788053424509028
command_prefix = "!"
use_embeds = true
auto_react = true
verbose = true
escalation_timeout = 300
